+++ cob_grasp_generation/scripts/generate_grasps_server.py
@@ -22,17 +22,15 @@ import moveit_msgs.msg
 import cob_grasp_generation.msg
 from cob_grasp_generation import or_grasp_generation, grasp_query_utils
 
-class CobGraspGenerationActionServer(object):
+class GenerateGraspsServer(object):
 
   def __init__(self):
     self.orgg = or_grasp_generation.ORGraspGeneration()
 
     self._as_generate = actionlib.SimpleActionServer('generate_grasps', cob_grasp_generation.msg.GenerateGraspsAction, execute_cb=self.generate_cb, auto_start = False)
     self._as_generate.start()
-    self._as_show = actionlib.SimpleActionServer('show_grasps', cob_grasp_generation.msg.ShowGraspsAction, execute_cb=self.show_cb, auto_start = False)
-    self._as_show.start()
 
-    print("CobGraspGenerationActionServer: actions started...")
+    print("GenerateGraspsServer: action started...")
 
   def generate_cb(self, goal):
     success = False
@@ -62,31 +60,7 @@ class CobGraspGenerationActionServer(object):
       self._as_generate.set_aborted(result)
 
 
-  def show_cb(self, goal):
-    success = False
-
-    rospy.loginfo('Show grasp %i for object %s using gripper_type %s' % (goal.grasp_id, goal.object_name, goal.gripper_type))
-
-    if grasp_query_utils.check_database(goal.object_name, goal.gripper_type):
-      self.orgg.setup_environment(goal.object_name, goal.gripper_type, viewer=True)
-      rospy.loginfo('Display Grasp. Object: %s | ID: %i' % (goal.object_name, goal.grasp_id))
-      self.orgg.show_grasp(goal.object_name, goal.gripper_type, goal.grasp_id, goal.sort_by_quality)
-      success = True
-    else:
-      rospy.logerr('GraspTable for Object %s does not exist!' % (goal.object_name))
-
-    result   = cob_grasp_generation.msg.ShowGraspsResult()
-    result.success = success
-
-    if success:
-      rospy.loginfo('Show: Succeeded')
-      self._as_show.set_succeeded(result)
-    else:
-      rospy.logwarn('Show: Failed')
-      self._as_show.set_aborted(result)
-
-
 if __name__ == '__main__':
-  rospy.init_node('generate_grasps')
-  CobGraspGenerationActionServer()
+  rospy.init_node('generate_grasps_server')
+  GenerateGraspsServer()
   rospy.spin()
